Revert "Bug 2032985 - Make PBrowser::RequestNativeKeyBindings more idiomatic. r=masayuki" for causing build bustages @ EnumSerializer
This reverts commit fdbf57f92ff8591af0b0e02d96b3908601a7324d.

+++ dom/ipc/BrowserChild.cpp
@@ -2396,7 +2396,8 @@ void BrowserChild::RequestEditCommands(NativeKeyBindingsType aType,
   // Don't send aEvent to the parent process directly because it'll be marked
   // as posted to remote process.
   WidgetKeyboardEvent localEvent(aEvent);
-  SendRequestNativeKeyBindings(aType, localEvent, &aCommands);
+  SendRequestNativeKeyBindings(static_cast<uint32_t>(aType), localEvent,
+                               &aCommands);
 }
 
 mozilla::ipc::IPCResult BrowserChild::RecvSynthesizedEventResponse(
